Fixed bug with boosting fields
When using the `BoostFunctions` parameter the wrong query/field
gets overwritten. When the boost field is decoded it
stored using numerical keys instead of the basename.
This key (which is now numerical)  is then later used by a call to
`array_search` which forces all values in the array to be converted
to a numerical value before comparing.

Example:
```php
array_search(0, ["title_t", "description_t"])
```
This will turn each value into a `0` due to type casting and it
will find a match for the first entry, it then returns they key `0` as
the result.

The first boost entry will have the key `0` which then triggers
this issue in `array_search` and it ends up overwriting the first
query-field no matter what it contains.

Example of `BoostFunctions` usage:
```php
$searchEngine->search($term, [
    'BoostFunctions' => [
        'fields' => ['another_title' => 5],
    ],
]);
```